Added Get-VhdSecurityPermission.ps1 and Set-VhdSecurityPermission.ps1
Get-VhdSecurityPermission.ps1 will enumerate the VHD's attached to a given guest VM and check if the VMID is added to the NTFS ACL with READ,WRITE, and SYNCHRONIZE permission.  If not, it should prompt the user to correct this.

Set-VhdSecurityPermission.ps1 allows you to explicitly define which VMID (defined by the -VMName parameter) to add to a specific VHD file (defined by the -Disk parameter).  This difference between this and Get-VhdSecurityPermission is it does not attempt to perform enumeration and gives you fine grained control of what VMID you're adding to a specified VHD file.